refactor(packages): 패키지 분리
