+++ src/main/java/racingcar/RacingCarApplication.java
@@ -1,5 +1,8 @@
 package racingcar;
 
+import racingcar.io.Input;
+import racingcar.io.Output;
+
 public class RacingCarApplication {
     public static void main(String[] args) {
         new RacingGame(new Input(), new Output()).run();
