+++ src/main/java/racingcar/io/Output.java
@@ -1,7 +1,9 @@
-package racingcar;
+package racingcar.io;
 
 import java.util.List;
 import java.util.stream.Collectors;
+import racingcar.domain.Car;
+import racingcar.vo.Name;
 
 public class Output {
     public void printCurrentPosition(List<Car> cars) {
