+++ src/main/java/racingcar/vo/Name.java
@@ -1,7 +1,6 @@
-package racingcar;
+package racingcar.vo;
 
 import java.util.Objects;
-import java.util.StringJoiner;
 
 public class Name {
     private static final int MAXIMUM_LENGTH = 5;
