+++ src/test/java/racingcar/InputNameHandlerTest.java
@@ -1,9 +1,9 @@
 package racingcar;
 
-import java.util.Arrays;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.vo.Name;
 import static org.assertj.core.api.Assertions.*;
 
 class InputNameHandlerTest {
