--- src/test/java/racingcar/RacingGameTest.java
@@ -1,11 +1,0 @@
-package racingcar;
-
-import java.util.List;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import static org.assertj.core.api.Assertions.*;
-
-public class RacingGameTest {
-
-
-}
