+++ src/test/java/racingcar/vo/NameTest.java
@@ -1,9 +1,10 @@
-package racingcar;
+package racingcar.vo;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.vo.Name;
 import static org.assertj.core.api.Assertions.*;
 
 public class NameTest {
